core: added CommandService to separate Command Execution from handling messages in general #145
