+++ src/main/java/de/voidtech/gerald/listeners/ReadyListener.java
@@ -3,11 +3,11 @@ package main.java.de.voidtech.gerald.listeners;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import main.java.de.voidtech.gerald.service.CommandService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import main.java.de.voidtech.gerald.service.GeraldConfig;
-import main.java.de.voidtech.gerald.service.MessageHandler;
 import main.java.de.voidtech.gerald.service.TwitchNotificationService;
 import net.dv8tion.jda.api.events.GenericEvent;
 import net.dv8tion.jda.api.events.ReadyEvent;
@@ -19,7 +19,7 @@ public class ReadyListener implements EventListener {
 	private static final Logger LOGGER = Logger.getLogger(GeraldConfig.class.getName());
 	
 	@Autowired
-	private MessageHandler messageHandler;
+	private CommandService cmdService;
 	
 	@Autowired
 	private TwitchNotificationService twitchService;
@@ -30,7 +30,7 @@ public class ReadyListener implements EventListener {
 			String clientName = ((ReadyEvent) event).getJDA().getSelfUser().getAsTag();
 			LOGGER.log(Level.INFO, "Coffee Machine is ready! Serving lattes as " + clientName);	
 			
-			messageHandler.loadAliases();	
+			cmdService.loadAliases();
 			twitchService.subscribeToAllStreamers();
 		}
 	}
